WIP: add various msg ids into the DB

+++ src/persistence/history.h
@@ -224,7 +224,7 @@ public:
     void addNewMessage(const ChatId& chatId, const QString& message, const ToxPk& sender,
                        const QDateTime& time, bool isDelivered, ExtensionSet extensions,
                        QString dispName, const std::function<void(RowId)>& insertIdCallback = {},
-                       const uint8_t hasIdType = 0);
+                       const int hasIdType = 0);
 
     void addPushtoken(const ToxPk& sender, const QString& pushtoken);
     QString getPushtoken(const ToxPk& friendPk);
